Recover seven field signednesses from their readers, and drop the puns they forced
Each of these fields is declared unsigned but every reader takes it signed, so
the reader's width is the load-bearing one and the declaration was wrong. With
the declaration corrected the *(s8*)& puns and (s8) casts at the use sites are
redundant and go away.

  ARWGeneratorSetup.spreadX/Y/Z (677)   u8  -> s8   negated as the low bound of
                                                    randomGetRange(-spread, spread),
                                                    and the sibling velocityX/Y/Z
                                                    at 0x1C are already s8
  KtrexfloorswitchState.graceTimer,     u8  -> s8   if ((state->graceTimer -= 1) < 0)
    .prevGraceTimer (593)
  KtrexState.pathCountdown,             u8  -> s8   decremented then tested <= 0 / < 1
    .hitCountdown (592)
  ArwSquadronState.volleyShotsRemaining u8  -> s8   decremented then tested <= 0
    (678)
  AlphaAnimatorState.completedCycles    u8  -> s8   all six readers cast (s8)
    (311)
  GridEntry.unk0B (gameui)              u8  -> s8   its only reader adds it to
                                                    GridEntry.x as a signed nudge;
                                                    named xOffset accordingly

All six units rebuild byte-identical (objcopy -R .comment; objdump -s -t -r).

+++ src/dlls/objects/592_KT_Rex/KT_Rex.c
@@ -492,7 +492,7 @@ int ktrex_stateHandlerA05(GameObject* obj, KTRexRuntime* runtime)
     if (ktrex_updateArenaPathProgress(runtime) != 0)
     {
         gKTRexState->pathCountdown -= 1;
-        if ((s8)gKTRexState->pathCountdown <= 0)
+        if (gKTRexState->pathCountdown <= 0)
         {
             pushLo = 2;
             if (Stack_IsFull(gKTRexState->stack) == 0)
